Refactor to a single EventsClient
